+++ NEWS.md
@@ -1,4 +1,12 @@
-# org 2019.02.21
+# org 2019.4.2
+
+- Allows for multiple code folders to be sourced using the argument `folders_to_be_sourced` (previously this was hardcoded as a folder called `code`)
+
+# org 2019.3.5
+
+- Removal of stringr and lubridate dependencies
+
+# org 2019.2.21
 
 - Submission to CRAN
 - Includes functions `AllowFileManipulationFromInitialiseProject` and `InitialiseProject`